Fixing this.plugin == null
