--- ModernFFA/src/main/java/org/plugin/modernffa/commands/ResetTask.java
@@ -1,7 +1,0 @@
-package org.plugin.modernffa.commands;
-
-import org.plugin.modernffa.ModernFFA;
-
-public interface ResetTask {
-    void runTaskTimer(ModernFFA plugin, long resetInterval, long resetInterval1);
-}
