--- ModernFFA/src/main/resources/README.MD
@@ -1,22 +1,0 @@
-We are pleased to introduce the FFA (Free-For-All) plugin compatible with Minecraft version 1.21. This sophisticated plugin supports an array of commands, enhancing the gameplay experience for both players and staff members. Below is a comprehensive overview of the commands:
-
-/ffa setspawn: Empowers staff members with the ffa.setspawn permission to designate the spawn location for players, either upon death or server entry.
-/ffa setarena <name>: Enables staff with the ffa.setarena permission to create and name specific arenas for player teleportation.
-/ffa setviparena: Allows staff with the ffa.setviparena permission to establish exclusive arenas accessible only to players with the requisite ranks or permissions.
-/ffa viparena: Permits players to join the VIP arena, contingent upon having the ffa.viparena permission.
-/ffa arena <name>: Facilitates player teleportation to pre-set arenas, displaying the arena's name prominently upon arrival. This command is universally accessible without any permission requirements.
-/ffa createkit <name>: Authorizes staff with the ffa.createkit permission to create kits that save the player’s current inventory items and effects.
-/ffa kit <name>: Allows players to equip pre-created kits.
-/ffa setpositionresetmap <name>: After executing this command in a specific arena, the user can delineate the arena boundaries using a wooden axe and finalize the setup with the ffa.setpointresetmap command.
-/ffa enableresetarena <name> <time>: This command resets a designated arena to its default state after a specified duration, replenishing or removing blocks, items, and entities as necessary. This process is optimized to avoid crashes or excessive resource consumption.
-In addition to these commands, the plugin features an automated health and hunger bar replenishment system upon a successful kill. When a player eliminates an opponent, the following messages appear:
-
-Messages supported:
-- Killer's POV: A green action-bar message: "You have killed <victim> with <number of hearts> left. You gained 20 hearts." (where <victim> and <number of hearts> are highlighted in red).
-- Victim's POV: A red action-bar message: "You have been killed by <killer>. They have <number of hearts> left." (with <killer> highlighted in orange).
-- Moreover, the plugin mitigates the issue of free-hitting by instituting a duel request system. When a player wishes to initiate combat, they must first punch their desired opponent, triggering a green chat message in the recipient's POV: "<hitter> wants to fight against you" (with <hitter> highlighted in orange). This message appears only once every 10 seconds to prevent spam. If the recipient reciprocates the hit, the duel commences.
-
-We trust that these features will significantly enhance your gaming experience. Should you require further assistance or have any queries, please do not hesitate to contact us.
-
-Happy gaming!
-
